packages/quri_parts/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning", "setuptools"]
build-backend = "poetry_dynamic_versioning.backend"



[tool.poetry]
name = "jijmodeling-transpiler-quantum-quri-parts"
version = "0.1.0rc2"
description = "JijModeling Transpiler for QURI parts"
authors = ["Jij Inc."]
readme = "README.md"
packages = [{include = "jijmodeling_transpiler_quantum"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
jijmodeling-transpiler-quantum = "*"
quri-parts = ">=0.11.0"
qulacs = ">=0.5.6"


[tool.poetry.group.dev.dependencies]
matplotlib = "^3.7.2"
jijmodeling-transpiler-quantum = {path = "../../", develop = true}
